add singlethreaded

+++ h2ox/forecast/tigge.py
@@ -56,6 +56,16 @@ def ingest_local_grib(grib_path, gcs_root, n_workers, zero_dt):
     ds = xr.open_dataset(grib_path, engine="cfgrib")
 
     start_dt = pd.to_datetime(ds["time"].min().values).to_pydatetime()
+    
+    time_offset = int((start_dt - zero_dt).days)
+    
+    # load the slices
+    slices = dask.array.core.slices_from_chunks(
+        dask.array.empty_like(ds_dask.to_array()).chunks
+    )  # null, time, lat, lon
+
+    # eliminate the slices which hit the boundary and only the first variable
+    slices = [s for s in slices if s[3].stop != 361 and s[0].start == 0]
 
     for variable in ds.data_vars.keys():
 
@@ -65,52 +75,66 @@ def ingest_local_grib(grib_path, gcs_root, n_workers, zero_dt):
 
         # transpose to correct shape
         data = np.transpose(data, [3, 2, 0, 1])
+        
+        if n_workers>1:
+            # do multiprocessing
 
-        logger.info("assigning to sharedemem")
-        # create the shared memory space
-        shm = SharedMemory(create=True, size=data.nbytes)
+            logger.info(f"assigning to sharedemem: {data.nbytes}")
+            # create the shared memory space
+            shm = SharedMemory(create=True, size=data.nbytes)
 
-        # create the shared array
-        data_shm = np.ndarray(data.shape, dtype=data.dtype, buffer=shm.buf)
+            # create the shared array
+            logger.info('Create shared array')
+            data_shm = np.ndarray(data.shape, dtype=data.dtype, buffer=shm.buf)
 
-        # write the data into shared mem
-        data_shm[:] = data[:]
+            # write the data into shared mem
+            data_shm[:] = data[:]
 
-        shm_spec = {"name": shm.name, "shape": data.shape, "dtype": data.dtype}
+            shm_spec = {"name": shm.name, "shape": data.shape, "dtype": data.dtype}
 
-        # load the slices
-        slices = dask.array.core.slices_from_chunks(
-            dask.array.empty_like(ds_dask.to_array()).chunks
-        )  # null, time, lat, lon
+            logger.info(f"Got {len(slices)} slices")
+            # prep for multiprocessing
+            chunk_worker = len(slices) // n_workers + 1
+            slices_rechunked = [
+                slices[chunk_worker * ii : chunk_worker * (ii + 1)]
+                for ii in range(n_workers)
+            ]
 
-        # eliminate the slices which hit the boundary and only the first variable
-        slices = [s for s in slices if s[3].stop != 361 and s[0].start == 0]
+            args = [
+                (shm_spec, variable, gcs_root, slices_rechunked[ii], time_offset, ii)
+                for ii in range(n_workers)
+            ]
 
-        logger.info(f"Got {len(slices)} slices")
-        # prep for multiprocessing
-        chunk_worker = len(slices) // n_workers + 1
-        slices_rechunked = [
-            slices[chunk_worker * ii : chunk_worker * (ii + 1)]
-            for ii in range(n_workers)
-        ]
+            logger.info("Calling MP Pool with sharedmem")
+            # single threaded first
+            # for arg in args:
+            #    sharedmem_worker(*arg)
 
-        time_offset = int((start_dt - zero_dt).days)
+            pool.starmap(sharedmem_worker, args)
 
-        args = [
-            (shm_spec, variable, gcs_root, slices_rechunked[ii], time_offset, ii)
-            for ii in range(n_workers)
-        ]
+            del data_shm  # Unnecessary; merely emphasizing the array is no longer used
+            shm.close()
+            shm.unlink()  # Free and release the shared memory block at the very end
+            
+        else:
+            
+            # open the zarr
+            store = gcsfs.GCSMap(root=gcs_root)
+            z = zarr.open(store)
 
-        logger.info("Calling MP Pool with sharedmem")
-        # single threaded first
-        # for arg in args:
-        #    sharedmem_worker(*arg)
+            # write each slice
+            for ii_s, s in enumerate(slices):
+                if ii_s % 100 == 0:
+                    logger.info(f"single_worker, ii_s:{ii_s}")
 
-        pool.starmap(sharedmem_worker, args)
-
-        del data_shm  # Unnecessary; merely emphasizing the array is no longer used
-        shm.close()
-        shm.unlink()  # Free and release the shared memory block at the very end
+                time_slice = s[1]
+                offset_slice = slice(s[1].start + time_offset, s[1].stop + time_offset)
+                step_slice = s[2]
+                lat_slice = s[3]
+                lon_slice = s[4]
+                z[variable][lon_slice, lat_slice, offset_slice, step_slice] = data[
+                    lon_slice, lat_slice, time_slice, step_slice
+                ]
 
 
 def sharedmem_worker(shm_spec, variable, gcs_path, slices, time_offset, worker_idx):
